cloud_functions/function-transmit/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-iot==2.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
idna==3.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==1.0.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.27
urllib3==1.26.7
